fix(review): full audit P0/P1 correctness and fail-closed guards
- api.py: recompute trusted RiskContext in /v1/risk/check and /v1/tfex/risk/check (do not trust client-supplied RiskContext) per invariant #3; add tfex_gateway_lock/settrade_feed_lock double-checked locking for SettradeTfexGateway/SettradeRealtimeFeed singletons

- service.py: remove legacy reusable live token bypass (hmac) — intent-bound approval mandatory; preserve broker truth in cancel_order merge (keep remote symbol/side/quantity/price, audit divergence) per #5

- broker/ccxt.py: reject UP/DOWN sides with ValueError, switch OrderRecord quantity/filled to float to preserve fractional crypto (BTC 0.015) per domain float migration; fix _status float handling

- domain.py: migrate OrderIntent/OrderRecord/FillRecord/Position quantity to float (gt0) for CCXT fractional support

- portfolio.py: Holding quantity float, can_sell/apply_fill float-aware with epsilon

- session_reconcile.py: expected/broker positions float with 1e-9 epsilon, PositionDiscrepancy float

- broker/prediction.py: quantity float mapping

- agent_os/engine.py: validate side in {buy,sell} and quantity>0, pass float quantity (no int truncation)

- prediction/live.py: outcome-specific token_id (market-YES/NO) to fix always-BUY bug, HMAC per-method/path, keep per-request auth

- frontend: ChartWidget hook violation (move useEffect before early return, placeholder after hooks); Navigation sessionStorage/localStorage mismatch and storage event; utils/api getBackendUrl runtime SSR fix (api:9569 inside container, zksato-api.zeaz.dev for prod)

- tests: adapt fake clients to market_id/token_id dual field

Validation: ruff check . passed, pytest 272 passed 1 skipped, docker health ok, verify_external_gates available:true

+++ src/zksato/domain.py
@@ -98,7 +98,7 @@ class Bar(BaseModel):
 class OrderIntent(BaseModel):
     symbol: str = Field(min_length=1, max_length=32)
     side: Side
-    quantity: int = Field(gt=0)
+    quantity: float = Field(gt=0)
     order_type: OrderType = OrderType.LIMIT
     price: float | None = Field(default=None, gt=0)
     stop_loss: float | None = Field(default=None, gt=0)
@@ -167,8 +167,8 @@ class OrderRecord(BaseModel):
     client_order_id: str | None = None
     symbol: str
     side: Side
-    quantity: int
-    filled_quantity: int = 0
+    quantity: float
+    filled_quantity: float = 0
     order_type: OrderType
     price: float | None
     average_fill_price: float | None = None
@@ -198,7 +198,7 @@ class FillRecord(BaseModel):
     broker_order_id: str | None = Field(default=None, max_length=128)
     symbol: str = Field(min_length=1, max_length=32)
     side: Side
-    quantity: int = Field(gt=0)
+    quantity: float = Field(gt=0)
     price: float = Field(gt=0)
     fee: float = Field(default=0.0, ge=0)
     timestamp: datetime = Field(default_factory=lambda: datetime.now(UTC))
@@ -226,7 +226,7 @@ class RiskEvaluation(BaseModel):
 
 class Position(BaseModel):
     symbol: str
-    quantity: int
+    quantity: float
     average_price: float
     market_price: float
     market_value: float
